Moved over old solutions

+++ Mastering_Critical_Skills/Skills/Section_04_Data_Types_and_Variables/01_Easy_Homework/homework_01_math_operations.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 
 auto main() -> int {
-    int a{}, b{};
+    double a{}, b{};
     std::cin >> a >> b;
 
     std::cout << a << " + " << b << " = " << a + b << std::endl;
